feat: better error handling + logic to automatically close browser
Link to Jira issue: https://qompa.atlassian.net/browse/QDA-384

This new implementation is here because we got the following error:
"page: Target page, context or browser has been closed" which I
identified is present at `browser.newPage` call, is not clear the reason
why that happened but at least now the system can automatically recover
in case something happens.

Every page instance has a timeout of 30s so if it fails it will throw,
unregister the page and eventually the "ripper" will close the browser
if something there is wrong. For that reason a retry logic might be
necessary in the client side.

If the browser is not initialized when a new request to get a sunat
token is received then a new one will be created and will be reused
across other requests until the ripper destroys it.

The reason there is a reaper to do that task is because I didn't
expected that but before the timeout implementation the pages were stuck
undefinitely because the browser was closed when is was actually being
used by the client.

Additionally I removed the old code to retrieve proposals data from the
SUNAT platform since it is already being extracted from the
sunat-scraper library and is unnecesary to have that logic in this API.